Final commit
This project was the most difficult one I have been presented with so
far. I had absolutely no idea where to begin with this, and if not for
some assistance from AI, I am not sure if I would have been able to
figure this one out. I was really reluctant to use AI to help with this
course, and this is the first time I have resorted to that avenue for
help. However, after speaking with my brother who works professionally
as a backend developer I was inclined to try it. He said that his boss
encourages their team to use it for help when needed, and that's good
enough for me. I will still aim to use it as little as possible in the
future, but I must admit it's a an incredibly capable tutor.